mac1 joins as a compute node; qwen2.5:14b is the pick on the real prompt
macOS nodes are built on the node itself (scripts/install-node-macos.sh) and
deployable with just deploy-node / update-node. Local Network privacy
blocks the unsigned agent from the LAN, so a cron watchdog keeps it up and
a loopback relay under /usr/bin/python3 carries the coordinator connection;
TLS still pins the coordinator's fingerprint end to end.

Also: the agent's hostname detection now works without /etc/hostname
(macOS registered as "unknown"), free_bytes_for builds on macOS, and
just update-coordinator ships a coordinator binary without touching state.

Real-prompt bench on mac1: qwen2.5:14b 7/7 native and prompt at 1.2 s a
command; qwen3:8b 7/7 native but 6/7 prompt. mac1 serves qwen2.5:14b.

+++ capabilities/llm/src/llama.rs
@@ -389,7 +389,10 @@ fn free_bytes_for(_path: &Path) -> u64 {
         let rc = unsafe { libc::statvfs(c_path.as_ptr(), buf.as_mut_ptr()) };
         if rc == 0 {
             let s = unsafe { buf.assume_init() };
-            return s.f_bavail.saturating_mul(s.f_bsize);
+            // statvfs field widths differ by platform: both u64 on Linux, but
+            // f_bavail is u32 on macOS. Widen both so one line builds on both.
+            #[allow(clippy::unnecessary_cast)]
+            return (s.f_bavail as u64).saturating_mul(s.f_bsize as u64);
         }
     }
     u64::MAX
